update fedora scripts

+++ linux/fedora/vim_and_python.sh
@@ -1,14 +1,17 @@
 #!/bin/bash
 # source: VIM and Python - a Match Made in Heaven, Real Python, https://realpython.com/blog/python/vim-and-python-a-match-made-in-heaven/
 
-# !!!! backup your .vimrc first !!!!
+# backup .vimrc
+if [ -f ~/.vimrc ] then
+  cp ~/.vimrc ~/.vimrc.bkp
+fi
 
-sudo dnf update -y
+#sudo dnf update -y  # let user handle update
 sudo dnf install vim git -y
 
 git clone https://github.com/gmarik/Vundle.vim.git ~/.vim/bundle/Vundle.vim
 
-cp ~/.vimrc ~/.vimrc.bkp
+
 echo "
 \" use spaces instead of tabs
 set tabstop=2
@@ -57,15 +60,14 @@ filetype plugin indent on    \" required
 " > ~/.vimrc
 vim +PluginInstall +qall
 
-sudo dnf update -y
 sudo dnf install python-devel -y # "Python.h"
 sudo dnf install python-pip -y 
 sudo dnf install python2-tkinter.x86_64 -y
 # tkinter is needed by matplotlib.pyplot
 
-# solve 'gcc error:/usr/lib/rpm/redhat/redhat-hardened-cc1: No such file or directory'
-sudo dnf install redhat-rpm-config -y
+## solve 'gcc error:/usr/lib/rpm/redhat/redhat-hardened-cc1: No such file or directory'
+#sudo dnf install redhat-rpm-config -y
 
-pip install --upgrade pip
+#pip install --upgrade pip  # bad idea to upgrade pip beyond system support
 pip install --user numpy scipy pandas matplotlib
 pip install --user lxml h5py scikit-image
